fix: 示例可直接运行 + Python 3.9 兼容 + 新增 3 个示例
- examples/_bootstrap.py: 路径引导，示例可从任意目录直接 python3 运行
- 6 个示例均加 import _bootstrap，修复 ModuleNotFoundError: No module named 'aui'
- 新增示例: form_curses(表单), gallery_curses(组件画廊), layout_ascii(布局)
- Python 3.9 兼容: curses 后端 list[...] 改回 typing.List (TList)，避免 3.9 不支持下标
- ascii 后端 _hline float 乘法修复
- curses 后端 _measure_child 统一单行压缩 + typing.List 遮蔽 components.List 修复
- 验证: venv 48 测试全绿; 系统 python3.9 从 examples 目录直接运行 OK

+++ src/aui/backends/ascii.py
@@ -121,8 +121,8 @@ class AsciiBackend:
             if 0 <= cx < self.width:
                 self._canvas[y][cx] = ch
 
-    def _hline(self, x: int, y: int, length: int) -> None:
-        self._put(x, y, "-" * max(0, length))
+    def _hline(self, x: int, y: int, length: float) -> None:
+        self._put(x, y, "-" * max(0, int(length)))
 
     def _box(self, x: int, y: int, label: str) -> None:
         text = " " + label + " "
